refactor: use metadata api and modular admission form
