+++ src/pages/Admission.jsx
@@ -8,6 +8,7 @@ import BASE_URL from '../config';
 import AdmissionForm from '../components/admission/AdmissionForm';
 import AdmissionCard from '../components/admission/AdmissionCard';
 import Modal from '../components/common/Modal';
+import { fetchMetadata } from '../utils/api';
 
 const Admission = () => {
   const initialForm = {
@@ -37,48 +38,16 @@ const Admission = () => {
 
   const institute_uuid = localStorage.getItem("institute_uuid");
 
-  const fetchCourses = async () => {
+  const fetchMeta = async () => {
     try {
-      const res = await axios.get(`${BASE_URL}/api/courses`);
-      setCourses(Array.isArray(res.data) ? res.data : []);
+      const data = await fetchMetadata(institute_uuid);
+      setCourses(Array.isArray(data.courses) ? data.courses : []);
+      setEducations(Array.isArray(data.educations) ? data.educations : []);
+      setExams(Array.isArray(data.exams) ? data.exams : []);
+      setBatches(Array.isArray(data.batches) ? data.batches : []);
+      setPaymentModes(Array.isArray(data.paymentModes) ? data.paymentModes : []);
     } catch {
-      toast.error('Failed to load courses');
-    }
-  };
-
-  const fetchEducations = async () => {
-    try {
-      const res = await axios.get(`${BASE_URL}/api/education`);
-      setEducations(Array.isArray(res.data) ? res.data : []);
-    } catch {
-      toast.error('Failed to load education options');
-    }
-  };
-
-  const fetchExams = async () => {
-    try {
-      const res = await axios.get(`${BASE_URL}/api/exams`);
-      setExams(Array.isArray(res.data) ? res.data : []);
-    } catch {
-      toast.error('Failed to load exam events');
-    }
-  };
-
-  const fetchBatches = async () => {
-    try {
-      const res = await axios.get(`${BASE_URL}/api/batches`);
-      setBatches(res.data || []);
-    } catch {
-      toast.error('Failed to load batches');
-    }
-  };
-
-  const fetchPaymentModes = async () => {
-    try {
-      const res = await axios.get(`${BASE_URL}/api/paymentmode`);
-      setPaymentModes(Array.isArray(res.data) ? res.data : []);
-    } catch {
-      toast.error('Failed to load payment modes');
+      toast.error('Failed to load form metadata');
     }
   };
 
@@ -186,11 +155,7 @@ const Admission = () => {
   };
 
   useEffect(() => {
-    fetchCourses();
-    fetchEducations();
-    fetchExams();
-    fetchBatches();
-    fetchPaymentModes();
+    fetchMeta();
     fetchAdmissions();
   }, []);
